rename repository package
