+++ core/src/main/java/GroceryFamily/GroceryElders/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package GroceryFamily.GroceryElders.repo;
+package GroceryFamily.GroceryElders.repository;
 
 import GroceryFamily.GroceryElders.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
